# api/openapi/catalog.yaml
openapi: 3.0.3
info:
  title: Catalog REST API
  description: REST API for Kubeflow Hub Catalog
  version: v1alpha1
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0"
servers:
  - url: "https://localhost:8080"
  - url: "http://localhost:8080"
paths:
...
  /api/model_catalog/v1alpha1/labels:
    summary: Path used to get the list of catalog labels.
    description: >-
      The REST endpoint/path used to list zero or more `CatalogLabel` entities.
    get:
      summary: List All CatalogLabels
      tags:
        - ModelCatalogService
      parameters:
        - $ref: "#/components/parameters/assetType"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/labelOrderBy"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/CatalogLabelListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findLabels
      description: Gets a list of all `CatalogLabel` entities.
  /api/model_catalog/v1alpha1/models:
    description: >-
      The REST endpoint/path used to list zero or more `CatalogModel` entities from all `CatalogSources`.
    get:
      summary: Search catalog models across sources.
      tags:
        - ModelCatalogService
      parameters:
        - name: recommendations
          in: query
          description: Sort models by lowest recommended latency using Pareto filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: targetRPS
          in: query
          description: Target requests per second for latency calculations
          required: false
          schema:
            type: integer
            format: int32
        - name: latencyProperty
          in: query
          description: Property name for latency metric
          required: false
          schema:
            type: string
            default: "ttft_p90"
        - name: rpsProperty
          in: query
          description: Property name for RPS metric
          required: false
          schema:
            type: string
            default: "requests_per_second"
        - name: hardwareCountProperty
          in: query
          description: Property name for hardware count
          required: false
          schema:
            type: string
            default: "hardware_count"
        - name: hardwareTypeProperty
          in: query
          description: Property name for hardware type grouping
          required: false
          schema:
            type: string
            default: "hardware_type"
        - name: source
          description: |-
            Filter models by source. Multiple values can be separated by commas
            to filter by multiple sources (OR logic). For example:
            ?source=huggingface,local will return models from either
            huggingface OR local sources.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          in: query
          required: false
        - name: q
          description: Free-form keyword search used to filter the response.
          schema:
            type: string
          in: query
          required: false
        - name: sourceLabel
          description: |-
            Filter models by the label associated with the source. Multiple
            values can be separated by commas. If one of the values is the
            string `null`, then models from every source without a label will
            be returned.
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
        - $ref: "#/components/parameters/filterQuery"
        - $ref: "#/components/parameters/pageSize"
        - name: orderBy
          style: form
          explode: true
          examples:
            orderBy:
              value: ID
          description: |-
            Specifies the order by criteria for listing entities.

            Supported values are:
            - CREATE_TIME
            - LAST_UPDATE_TIME
            - ID
            - NAME
            - ACCURACY

            Defaults to `NAME`.

            The `ACCURACY` sort will sort by the `overall_average` property in any linked metrics artifact.

            In addition, models can be sorted by properties. For example:
            - `provider.string_value` sorts by provider name
            - `artifacts.ifeval.double_value` sorts by the min/max value a property called ifeval across all associated artifacts
          schema:
            $ref: "#/components/schemas/OrderByField"
          in: query
          required: false
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/CatalogModelListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findModels
  /api/model_catalog/v1alpha1/models/export:
    description: >-
      Export catalog models as a CSV file or preview the export as JSON.
    get:
      summary: Export catalog models as CSV or preview export.
      tags:
        - ModelCatalogService
      parameters:
        - name: dryRun
          in: query
          description: |-
            When true, returns a JSON preview containing the total count, column
            headers, and a paginated list of models that would be exported.
            When false or omitted, returns a CSV file download.
          required: false
          schema:
            type: boolean
            default: false
        - name: id
          in: query
          description: |-
            Export only models with these IDs. Repeatable for multiple IDs.
            Mutually exclusive with `excludeId` and filter parameters
            (`source`, `q`, `sourceLabel`, `filterQuery`).
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: excludeId
          in: query
          description: |-
            Exclude models with these IDs from the export. Repeatable for
            multiple IDs. Mutually exclusive with `id`.
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: source
          description: |-
            Filter models by source. Multiple values can be separated by commas
            to filter by multiple sources (OR logic).
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
          in: query
          required: false
        - name: q
          description: Free-form keyword search used to filter the response.
          schema:
            type: string
          in: query
          required: false
        - name: sourceLabel
          description: |-
            Filter models by the label associated with the source. Multiple
            values can be separated by commas. If one of the values is the
            string `null`, then models from every source without a label will
            be returned.
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
        - $ref: "#/components/parameters/filterQuery"
        - $ref: "#/components/parameters/pageSize"
        - name: orderBy
          style: form
          explode: true
          description: |-
            Specifies the order by criteria for listing entities.

            Supported values are:
            - CREATE_TIME
            - LAST_UPDATE_TIME
            - ID
            - NAME

            Defaults to `NAME`.
          schema:
            $ref: "#/components/schemas/OrderByField"
          in: query
          required: false
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          description: |-
            CSV file download (when dryRun is false or omitted) or JSON
            preview (when dryRun is true).
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: "#/components/schemas/ExportDryRunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: exportModels
  /api/model_catalog/v1alpha1/models/filter_options:
    description: Lists options for `filterQuery` when listing models.
    get:
      summary: Lists fields and available options that can be used in `filterQuery` on the list models endpoint.
      tags:
        - ModelCatalogService
      responses:
        "200":
          $ref: "#/components/responses/FilterOptionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findModelsFilterOptions
  /api/model_catalog/v1alpha1/sources:
    summary: Path used to get the list of catalog sources.
    description: >-
      The REST endpoint/path used to list zero or more `CatalogSource` entities.
    get:
      summary: List All CatalogSources
      tags:
        - ModelCatalogService
      parameters:
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/assetType"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/orderBy"
        - $ref: "#/components/parameters/sortOrder"
        - $ref: "#/components/parameters/nextPageToken"
      responses:
        "200":
          $ref: "#/components/responses/CatalogSourceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      operationId: findSources
      description: Gets a list of all `CatalogSource` entities.
...
components:
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
      description: Bad Request parameters
...
  parameters:
...
    labelOrderBy:
      style: form
      explode: true
      examples:
        labelOrderBy:
          value: name
      name: orderBy
      description: |
        Specifies the key to order catalog labels by. You can provide any string key
        that may exist in the label maps. Labels that contain the specified key will
        be sorted by that key's value. Labels that don't contain the key will maintain
        their original order and appear after labels that do contain the key.
      schema:
        type: string
      in: query
      required: false
...
    assetType:
      name: assetType
      description: Filter by asset type.
      schema:
        $ref: "#/components/schemas/CatalogAssetType"
      in: query
      required: false
...
    pageSize:
      examples:
        pageSize:
          value: "100"
      name: pageSize
      description: Number of entities in each page.
      schema:
        type: string
      in: query
      required: false
...
    sortOrder:
      style: form
      explode: true
      examples:
        sortOrder:
          value: DESC
      name: sortOrder
      description: "Specifies the sort order for listing entities, defaults to ASC."
      schema:
        $ref: "#/components/schemas/SortOrder"
      in: query
      required: false
...
  securitySchemes:
    Bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Bearer JWT scheme
security:
  - Bearer: []
...